requirements.txt
lpips==0.1.4
numpy==1.19.5
pystrum==0.4
PyYAML==6.0.1
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tqdm==4.64.1
